refactor: YoutubeVideoSourceのディレクトリ整理
Fixes #609
