--- src/resolvers/Mutation/registerVideoFromYoutube/prisma.ts
@@ -1,151 +1,0 @@
-import {
-  SemitagEventType,
-  Video,
-  VideoEventType,
-  VideoTagEventType,
-  VideoThumbnailEventType,
-  VideoTitleEventType,
-  YoutubeVideoSourceEventType,
-} from "@prisma/client";
-import { ulid } from "ulid";
-
-import { ok, Result } from "../../../utils/Result.js";
-import { ResolverDeps } from "../../types.js";
-
-export const register = async (
-  prisma: ResolverDeps["prisma"],
-  {
-    authUserId,
-    primaryTitle,
-    extraTitles,
-    primaryThumbnail,
-    tagIds,
-    semitagNames,
-    sourceIds,
-  }: {
-    authUserId: string;
-    primaryTitle: string;
-    extraTitles: string[];
-    primaryThumbnail: string;
-    tagIds: string[];
-    semitagNames: string[];
-    sourceIds: string[];
-  }
-): Promise<Result<{ type: "NO_TAG"; tagId: string } | { type: "INTERNAL_SERVER_ERROR"; error: unknown }, Video>> => {
-  const videoId = ulid();
-  const dataTitles = [
-    { id: ulid(), title: primaryTitle, isPrimary: true },
-    ...extraTitles.map((extraTitle) => ({
-      id: ulid(),
-      title: extraTitle,
-      isPrimary: false,
-    })),
-  ];
-  const dataThumbnails = [
-    {
-      id: ulid(),
-      imageUrl: primaryThumbnail,
-      isPrimary: true,
-    },
-  ];
-  const dataTags = tagIds.map((tagId) => ({
-    id: ulid(),
-    tagId,
-  }));
-  const dataSemitags = semitagNames.map((name) => ({
-    id: ulid(),
-    name,
-    isChecked: false,
-  }));
-  const dataYoutubeSources = sourceIds.map((sourceId) => ({
-    id: ulid(),
-    sourceId,
-  }));
-
-  const [video] = await prisma.$transaction([
-    prisma.video.create({
-      data: {
-        id: videoId,
-        titles: { createMany: { data: dataTitles } },
-        thumbnails: { createMany: { data: dataThumbnails } },
-        tags: { createMany: { data: dataTags } },
-        semitags: { createMany: { data: dataSemitags } },
-        youtubeSources: { createMany: { data: dataYoutubeSources } },
-      },
-    }),
-    prisma.videoEvent.createMany({
-      data: [
-        {
-          userId: authUserId,
-          videoId,
-          type: VideoEventType.REGISTER,
-          payload: {},
-        },
-      ],
-    }),
-    prisma.videoTitleEvent.createMany({
-      data: [
-        ...dataTitles.map(({ id }) => ({
-          userId: authUserId,
-          videoTitleId: id,
-          type: VideoTitleEventType.CREATE,
-          payload: {},
-        })),
-        {
-          userId: authUserId,
-          videoTitleId: dataTitles[0].id,
-          type: VideoTitleEventType.SET_PRIMARY,
-          payload: {},
-        },
-      ],
-    }),
-    prisma.videoThumbnailEvent.createMany({
-      data: [
-        ...dataThumbnails.map(({ id }) => ({
-          userId: authUserId,
-          videoThumbnailId: id,
-          type: VideoThumbnailEventType.CREATE,
-          payload: {},
-        })),
-        {
-          userId: authUserId,
-          videoThumbnailId: dataThumbnails[0].id,
-          type: VideoThumbnailEventType.SET_PRIMARY,
-          payload: {},
-        },
-      ],
-    }),
-    prisma.videoTagEvent.createMany({
-      data: [
-        ...dataTags.map(({ id }) => ({
-          userId: authUserId,
-          videoTagId: id,
-          type: VideoTagEventType.ATTACH,
-          payload: {},
-        })),
-      ],
-    }),
-    prisma.semitagEvent.createMany({
-      data: [
-        ...dataSemitags.map(({ id }) => ({
-          userId: authUserId,
-          semitagId: id,
-          type: SemitagEventType.ATTACH,
-          payload: {},
-        })),
-      ],
-    }),
-    prisma.youtubeVideoSourceEvent.createMany({
-      data: [
-        ...dataYoutubeSources.map(({ id }) => ({
-          userId: authUserId,
-          sourceId: id,
-          type: YoutubeVideoSourceEventType.CREATE,
-          payload: {},
-        })),
-      ],
-    }),
-  ]);
-
-  return ok(video);
-};
