+++ src/resolvers/Mutation/requestYoutubeRegistration/resolver.ts
@@ -1,11 +1,11 @@
 import { GraphQLError } from "graphql";
 
 import { isErr } from "../../../utils/Result.js";
+import { YoutubeVideoSourceDTO } from "../../../YoutubeVideoSource/dto.js";
 import { MutationResolvers, ResolversTypes } from "../../graphql.js";
 import { parseGqlID2 } from "../../id.js";
 import { ResolverDeps } from "../../types.js";
 import { YoutubeRegistrationRequestModel } from "../../YoutubeRegistrationRequest/model.js";
-import { YoutubeVideoSourceModel } from "../../YoutubeVideoSource/model.js";
 import { requestRegistration } from "./request.js";
 
 export const resolverRequestYoutubeRegistration = ({ prisma, logger }: Pick<ResolverDeps, "prisma" | "logger">) =>
@@ -46,7 +46,7 @@ export const resolverRequestYoutubeRegistration = ({ prisma, logger }: Pick<Reso
         case "VIDEO_ALREADY_REGISTERED":
           return {
             __typename: "RequestYoutubeRegistrationVideoAlreadyRegisteredError",
-            source: YoutubeVideoSourceModel.fromPrisma(result.error.source),
+            source: YoutubeVideoSourceDTO.fromPrisma(result.error.source),
           } satisfies ResolversTypes["RequestYoutubeRegistrationReturnUnion"];
         case "INTERNAL_SERVER_ERROR":
           logger.error({ error: result.error.error, path: info.path }, "Something error happens");
