--- src/resolvers/Query/findYoutubeVideoSource/findYoutubeVideoSource.graphql
@@ -1,3 +1,0 @@
-type Query {
-  findYoutubeVideoSource(sourceId: String): YoutubeVideoSource
-}
