--- src/resolvers/Query/getYoutubeVideoSource/getYoutubeVideoSource.graphql
@@ -1,3 +1,0 @@
-type Query {
-  getYoutubeVideoSource(id: ID!): YoutubeVideoSource!
-}
