+++ src/resolvers/Query/index.ts
@@ -1,6 +1,8 @@
 /* eslint sort-keys: [2, "asc", {caseSensitive: false}] */
 
 import { getSoundcloudVideoSource } from "../../SoundcloudVideoSource/getSoundcloudVideoSource.resolver.js";
+import { resolverFindYoutubeVideoSource } from "../../YoutubeVideoSource/findYoutubeVideoSource.resolver.js";
+import { getYoutubeVideoSource } from "../../YoutubeVideoSource/getYoutubeVideoSource.resolver.js";
 import { type Resolvers } from "../graphql.js";
 import { ResolverDeps } from "../types.js";
 import { fetchNicovideo } from "./fetchNicovideo/fetchNicovideo.js";
@@ -20,7 +22,6 @@ import { findUser } from "./findUser/findUser.js";
 import { findVideo } from "./findVideo/findVideos.js";
 import { findVideos } from "./findVideos/findVideos.js";
 import { resolverFindYoutubeRegistrationRequest } from "./findYoutubeRegistrationRequest/resolver.js";
-import { resolverFindYoutubeVideoSource } from "./findYoutubeVideoSource/resolver.js";
 import { resolverGetAllCategoryTag } from "./getAllCategoryTag/resolver.js";
 import { resolverGetAllTypeCategoryTag } from "./getAllTypeCategoryTag/resolver.js";
 import { getMylist } from "./getMylist/resolver.js";
@@ -33,7 +34,6 @@ import { getTag } from "./getTag/resolver.js";
 import { getUser } from "./getUser/resolver.js";
 import { getVideo } from "./getVideo/resolver.js";
 import { resolverGetYoutubeRegistrationRequest } from "./getYoutubeRegistrationRequest/resolver.js";
-import { getYoutubeVideoSource } from "./getYoutubeVideoSource/resolver.js";
 import { resolverNotifications } from "./notifications/resolver.js";
 import { searchTags } from "./searchTags/resolver.js";
 import { searchVideos } from "./searchVideos/resolver.js";
