--- src/resolvers/YoutubeVideoSource/YoutubeVideoSource.graphql
@@ -1,12 +1,0 @@
-type YoutubeVideoSource implements Node & VideoSource {
-  id: ID!
-  sourceId: String!
-  video: Video!
-
-  url: String!
-  embedUrl: String!
-}
-
-input YoutubeVideoSourceEventsOrderBy {
-  createdAt: SortOrder
-}
