--- src/resolvers/YoutubeVideoSource/events/events.graphql
@@ -1,11 +1,0 @@
-type YoutubeVideoSource {
-  events(
-    first: Int
-    after: String
-
-    last: Int
-    before: String
-
-    orderBy: YoutubeVideoSourceEventsOrderBy!
-  ): YoutubeVideoSourceEventConnection!
-}
