--- src/resolvers/YoutubeVideoSourceEvent/YoutubeVideoSource.graphql
@@ -1,17 +1,0 @@
-interface YoutubeVideoSourceEvent implements Node & Event {
-  id: ID!
-  series: String!
-  createdAt: DateTime!
-  user: User!
-
-  source: YoutubeVideoSource!
-}
-
-type YoutubeVideoSourceCreateEvent implements Node & Event & YoutubeVideoSourceEvent {
-  id: ID!
-  series: String!
-  createdAt: DateTime!
-  user: User!
-
-  source: YoutubeVideoSource!
-}
