--- src/resolvers/YoutubeVideoSourceEventConnection/YoutubeVideoSourceEventConnection.graphql
@@ -1,11 +1,0 @@
-type YoutubeVideoSourceEventConnection {
-  nodes: [YoutubeVideoSourceEvent!]!
-  edges: [YoutubeVideoSourceEventEdge!]!
-  pageInfo: PageInfo!
-  totalCount: Int!
-}
-
-type YoutubeVideoSourceEventEdge {
-  cursor: String!
-  node: YoutubeVideoSourceEvent!
-}
